+++ src/stave.ts
@@ -14,7 +14,9 @@ import { Clef } from './clef';
 import { KeySignature } from './keysignature';
 import { TimeSignature } from './timesignature';
 import { Volta } from './stavevolta';
-import { Bounds, StaveTempoOptions, FontInfo } from './types/common';
+import { Bounds, FontInfo } from './types/common';
+import { StaveTempoOptions } from './stavetempo';
+
 export interface StaveLineConfig {
   visible: boolean;
 }
